@@ -1,9 +1,7 @@
 package com.example.popularity.utils;
 
-import android.support.v4.app.Fragment;
+
+import androidx.fragment.app.Fragment;
 
 public class BaseFragment extends Fragment {
-
-
-
 }
