@@ -1,4 +1,4 @@
-package com.example.popularity.Utils;
+package com.example.popularity.utils;
 
 public class Configs {
     public static final String FACEBOOK_ROUT = "https://graph.facebook.com/";
