@@ -6,7 +6,6 @@ import com.example.popularity.R;
 import com.facebook.FacebookSdk;
 import com.facebook.appevents.AppEventsLogger;
 
-import uk.co.chrisjenx.calligraphy.CalligraphyConfig;
 
 public class MyApp extends Application {
     @Override
@@ -15,7 +14,7 @@ public class MyApp extends Application {
 
         FacebookSdk.sdkInitialize(getApplicationContext());
         AppEventsLogger.activateApp(this);
-        CalligraphyConfig.initDefault(new CalligraphyConfig.Builder().setDefaultFontPath("fonts/Samim_FD.ttf").build());
+       // CalligraphyConfig.initDefault(new CalligraphyConfig.Builder().setDefaultFontPath("fonts/Samim_FD.ttf").build());
 
     }
 }
